+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+- 2026-??-?? 2.3.22
+    - added Reading.protocol
+    - XS
+        - disable detector timeout during area scan
+        - adding Auto-Raman to BLEDevice
+        - only use one ACQUIRE to start Area Scan readout
+        - send extra ACQUIRE on occasional timeout
+- 2026-04-09 2.3.21
+    - XS Area Scan
+        - fixed line increment
+        - disable area scan while changing line increment
+    - XS BLE
+        - move firmware revisions to SpectrometerSettings
+        - fix gain
+        - fix laser enable
+        - fixed heartbeat
+        - working on LASER_STATE and BATTERY_STATE
+- 2026-04-09 2.3.20
+    - XS BLE
+        - fix None in utils.to_hex
+        - only parse first 8 EEPROM fields until BLE FW updated to support 9 
+        - stop checking for old FX2 code string in corrupted EEPROMs
+        - support V2.EVT prefix
+        - add new ACQUIRE status codes
+        - add new SPECTRA Characteristic
+        - read as many EEPROM bytes per page as supported by Peripheral
 - 2026-04-06 2.3.18
     - replaced WrapperWorker.is_FOO with .class_name
     - XL / Andor